==> vector-tokenizer/vectorize/1 loading_data.ipynb <==
## Loading data for the examples

We use an open Kaggle dataset for the examples in this project. Specifically, we utilize the [Delhi 5-Minute Electricity Demand for Forecasting dataset](https://www.kaggle.com/datasets/yug201/delhi-5-minute-electricity-demand-for-forecasting), which provides high-frequency electricity demand data for Delhi. This dataset is ideal for demonstrating time series analysis, forecasting techniques, and anomaly detection. By applying various modeling approaches, including PCA and machine learning, we explore patterns in electricity consumption and showcase the capabilities of TimeScape in extracting insights from temporal data.

import pandas as pd
import kagglehub

# Download latest version
path = kagglehub.dataset_download("yug201/delhi-5-minute-electricity-demand-for-forecasting")

print("Path to dataset files:", path)

## Dataset
Lets open the CSV file and load it into a dataframe.

data = pd.read_csv(path + "/powerdemand_5min_2021_to_2024_with weather.csv")
data.head()

The dataset consists of the following columns:

* datetime: Timestamp of the observation
* Power demand: Electricity demand (in kW) recorded every 5 minutes.
* temp: Temperature (°C).
* dwpt: Dew point temperature (°C).
* rhum: Relative humidity (%).
* wdir: Wind direction (degrees).
* wspd: Wind speed (m/s).
* pres: Atmospheric pressure (hPa).
* year, month, day, hour, minute: Breakdown of the timestamp for easy time-series analysis.

data.datetime = pd.to_datetime(data['datetime'])
data = data.set_index("datetime", drop=True)
data = data.drop(columns=["Unnamed: 0", "minute"])

We see that the dataset contains about 400k rows, a nice base for showcasing the performance of library functions.

data.describe()

Now we can store the dataset for convenience and later retrieval:

data.to_parquet("dehli.parquet")
